software/src/requirements.txt
pandas==2.2.3
python-Levenshtein==0.27.1